move Register page
