+++ src/index.js
@@ -1,19 +1,19 @@
-import ReactDOM from 'react-dom';
-import promiseFinally from 'promise.prototype.finally';
-import React from 'react';
-import { HashRouter } from 'react-router-dom';
-import { useStrict } from 'mobx';
-import { Provider } from 'mobx-react';
+import ReactDOM from "react-dom";
+import promiseFinally from "promise.prototype.finally";
+import React from "react";
+import { HashRouter } from "react-router-dom";
+import { useStrict } from "mobx";
+import { Provider } from "mobx-react";
 
-import App from './components/App';
+import App from "./App";
 
-import articlesStore from './stores/articlesStore';
-import commentsStore from './stores/commentsStore';
-import authStore from './stores/authStore';
-import commonStore from './stores/commonStore';
-import editorStore from './stores/editorStore';
-import userStore from './stores/userStore';
-import profileStore from './stores/profileStore';
+import articlesStore from "./stores/articlesStore";
+import commentsStore from "./stores/commentsStore";
+import authStore from "./stores/authStore";
+import commonStore from "./stores/commonStore";
+import editorStore from "./stores/editorStore";
+import userStore from "./stores/userStore";
+import profileStore from "./stores/profileStore";
 
 const stores = {
   articlesStore,
@@ -22,7 +22,7 @@ const stores = {
   commonStore,
   editorStore,
   userStore,
-  profileStore,
+  profileStore
 };
 
 // For easier debugging
@@ -31,10 +31,11 @@ window._____APP_STATE_____ = stores;
 promiseFinally.shim();
 useStrict(true);
 
-ReactDOM.render((
+ReactDOM.render(
   <Provider {...stores}>
     <HashRouter>
       <App />
     </HashRouter>
-  </Provider>
-), document.getElementById('root'));
+  </Provider>,
+  document.getElementById("root")
+);
