+++ src/pages/Register.js
@@ -1,12 +1,11 @@
-import { Link } from 'react-router-dom';
-import ListErrors from './ListErrors';
-import React from 'react';
-import { inject, observer } from 'mobx-react';
+import { Link } from "react-router-dom";
+import ListErrors from "components/ListErrors";
+import React from "react";
+import { inject, observer } from "mobx-react";
 
-@inject('authStore')
+@inject("authStore")
 @observer
 export default class Register extends React.Component {
-
   componentWillUnmount() {
     this.props.authStore.reset();
   }
@@ -14,10 +13,9 @@ export default class Register extends React.Component {
   handleUsernameChange = e => this.props.authStore.setUsername(e.target.value);
   handleEmailChange = e => this.props.authStore.setEmail(e.target.value);
   handlePasswordChange = e => this.props.authStore.setPassword(e.target.value);
-  handleSubmitForm = (e) => {
+  handleSubmitForm = e => {
     e.preventDefault();
-    this.props.authStore.register()
-      .then(() => this.props.history.replace('/'));
+    this.props.authStore.register().then(() => this.props.history.replace("/"));
   };
 
   render() {
@@ -27,20 +25,16 @@ export default class Register extends React.Component {
       <div className="auth-page">
         <div className="container page">
           <div className="row">
-
             <div className="col-md-6 offset-md-3 col-xs-12">
               <h1 className="text-xs-center">Sign Up</h1>
               <p className="text-xs-center">
-                <Link to="login">
-                  Have an account?
-                </Link>
+                <Link to="login">Have an account?</Link>
               </p>
 
               <ListErrors errors={errors} />
 
               <form onSubmit={this.handleSubmitForm}>
                 <fieldset>
-
                   <fieldset className="form-group">
                     <input
                       className="form-control form-control-lg"
@@ -78,11 +72,9 @@ export default class Register extends React.Component {
                   >
                     Sign in
                   </button>
-
                 </fieldset>
               </form>
             </div>
-
           </div>
         </div>
       </div>
